refactor(installer-ui): split app rendering helpers
