+++ docs/architecture/maintainability-baseline.txt
@@ -3,4 +3,3 @@
 813 crates/surge-cli/src/commands/install/mod.rs
 1860 crates/surge-cli/src/commands/install/remote.rs
 782 crates/surge-core/src/releases/delta.rs
-760 crates/surge-installer-ui/src/app.rs
